pyproject.toml
[tool.poetry]
name = "textgpt"
version = "0.1.0"
description = "pdf-to-gpt"
authors = ["sunilkumardash9 <sunilkumardash9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
gradio = "^3.27.0"
openai = "^0.27.4"
gradio-tools = "^0.0.5"
python-dotenv = "^1.0.0"
langchain = "^0.0.148"
chromadb = "^0.3.21"
tiktoken = "^0.3.3"
pypdf = "^3.8.1"
pymupdf = "^1.22.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.22.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
